REFACTOR: #60 상태 간소화 및 useInfiniteQuery를 사용하여 에러 해결
이전까지는 여러 상태를 의존하여 action을 결정하는 로직들이 너무 많이 있었는데,
이런 로직을 최대한 없애고, 상태를 간소화하였다. pagination의 경우, 기존에는
pageNum이라는 상태변수를 기반으로, 이 변수를 변경하면 다음 페이지 요청이 보내지도록 했는데,
useInfiniteQuery를 사용하면서 해당 상태변수를 삭제하였고, 이에 따라 상태를 기반으로 한
페이지 요청도 제거하였다.
이전과 같이 필터검색 요청과 pagination 요청은 분리되어 있는 상태인데, 이전에는 이 두 요청의 결과를
하나의 상태변수에 저장하여 상황에 따라 둘 중 하나의 데이터를 저장하도록 했는데, 이 또한
너무 상태 dependent한 로직이라 삭제하였다. 해당 데이터들을 상태로 저장하지 않아도, useInfiniteQuery의
결과로 받은 데이터의 page 개수를 기반으로 어떤 데이터를 display할지 결정해줘도 충분하다.
에러의 명확한 원인은 알아내지 못했으나, 너무 상태 의존적인 움직임을 남발하면 문제가 발생할 수 있다는 것을
알게되었다.
